pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ot_experiment_framework"
version = "0.1.0"
description = "Framework for generating and running OceanTracker particle-tracking experiments"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
]

[project.scripts]
ot-generate = "framework.generate:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["framework*"]
